refactor: 移动pool到pipeline包
